web/new_naraapi/naraapi_backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
anthropic==0.23.1
openai==1.23.2
python-dotenv==1.0.1
httpx==0.27.0

